+++ rh/forms.py
@@ -1,139 +1,7 @@
-import json
 from .models import *
 
 from django import forms
-from django.contrib.auth.forms import UserCreationForm, PasswordResetForm, SetPasswordForm, PasswordChangeForm
-from django.contrib.auth import get_user_model, password_validation
 from django.urls import reverse_lazy
-
-User = get_user_model()
-
-
-class RegisterForm(UserCreationForm):
-    """Subclass User Registration form"""
-
-    class Meta:
-        model = User
-        fields = ['name', 'username', 'email', 'password1', 'password2', 'organization']
-
-    def clean_username(self):
-        """check if username already exists"""
-        username = self.cleaned_data.get("username")
-        qs = User.objects.filter(username__iexact=username)
-        if qs.exists():
-            raise forms.ValidationError(
-                f"{username} is an invalid username, please pick another."
-            )
-        return username
-
-    def clean_email(self):
-        """check if email already exists"""
-        email = self.cleaned_data.get("email")
-        qs = User.objects.filter(email__iexact=email)
-        if qs.exists():
-            raise forms.ValidationError(
-                f"{email} is an invalid email address, please pick another."
-            )
-        return email
-
-
-class UserPasswordChangeForm(PasswordChangeForm):
-    """Subclass Password change form to remove labels and classes"""
-
-    def __init__(self, *args, **kwargs):
-        """Call super"""
-        super(UserPasswordChangeForm, self).__init__(*args, **kwargs)
-
-    old_password = forms.CharField(
-        label='',
-        strip=False,
-        widget=forms.PasswordInput(
-            attrs={
-                "autocomplete": "current-password",
-                "autofocus": True,
-                'class': 'form-control',
-                'placeholder': 'Old Password',
-                'type': 'password',
-                'name': 'old_password',
-
-            }
-        ),
-    )
-
-    new_password1 = forms.CharField(label='',
-                                    help_text=password_validation.password_validators_help_text_html(),
-                                    widget=forms.PasswordInput(
-                                        attrs={
-                                            'class': 'form-control',
-                                            'placeholder': 'New Password',
-                                            'type': 'password',
-                                            'name': 'new_password1',
-                                            'autocomplete': 'new-password'
-                                        }
-                                    ),
-                                    )
-    new_password2 = forms.CharField(label='',
-                                    widget=forms.PasswordInput(
-                                        attrs={
-                                            'class': 'form-control',
-                                            'placeholder': 'Confirm New Password',
-                                            'type': 'password',
-                                            'name': 'new_password2',
-                                            "autocomplete": "new-password"
-                                        }
-                                    ),
-                                    )
-
-
-class UserSetPasswordForm(SetPasswordForm):
-    """Override password reset form to add styling to the input fields"""
-
-    def __init__(self, *args, **kwargs):
-        """Call super"""
-        super(UserSetPasswordForm, self).__init__(*args, **kwargs)
-
-    new_password1 = forms.CharField(label='',
-                                    help_text=password_validation.password_validators_help_text_html(),
-                                    widget=forms.PasswordInput(
-                                        attrs={
-                                            'class': 'form-control',
-                                            'placeholder': 'New Password',
-                                            'type': 'password',
-                                            'name': 'new_password1',
-                                            'autocomplete': 'new-password'
-                                        }
-                                    ),
-                                    )
-    new_password2 = forms.CharField(label='',
-                                    widget=forms.PasswordInput(
-                                        attrs={
-                                            'class': 'form-control',
-                                            'placeholder': 'Confirm New Password',
-                                            'type': 'password',
-                                            'name': 'new_password2',
-                                            "autocomplete": "new-password"
-                                        }
-                                    ),
-                                    )
-
-
-class UserPasswordResetForm(PasswordResetForm):
-    """Override password reset form to add styling to the input fields"""
-
-    def __init__(self, *args, **kwargs):
-        """Call super"""
-        super(UserPasswordResetForm, self).__init__(*args, **kwargs)
-
-    email = forms.EmailField(label='',
-                             widget=forms.EmailInput(
-                                 attrs={
-                                     'class': 'form-control',
-                                     'placeholder': 'e.g abc@gmail.com',
-                                     'type': 'email',
-                                     'name': 'email'
-                                 }
-                             ),
-                             )
 
 
 class FieldHandler:
@@ -238,21 +106,3 @@ class ProjectForm(forms.ModelForm):
             'start_date': forms.widgets.DateInput(attrs={'type': 'date'}),
             'end_date': forms.widgets.DateInput(attrs={'type': 'date'}),
         }    
-
-
-class WarehouseLocationForm(forms.ModelForm):
-    class Meta:
-        model = WarehouseLocation
-        fields = "__all__"
-    
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['province'].queryset = Location.objects.filter(type='Province')
-        self.fields['district'].queryset = Location.objects.filter(type='District')
-
-
-class StockLocationDetailsForm(forms.ModelForm):
-    class Meta:
-        model = StockLocationDetails
-        fields = "__all__"
-
