+++ settings.py
@@ -23,6 +23,8 @@ ALLOWED_HOSTS = ['dev.reporthub.immap.org', '127.0.0.1']
 # Application definition
 INSTALLED_APPS = [
     'rh.apps.RhConfig',
+    'accounts.apps.AccountsConfig',
+    'stock.apps.StockConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -107,7 +109,7 @@ PASSWORD_HASHERS = [
     'django.contrib.auth.hashers.BCryptPasswordHasher',
 ]
 
-AUTH_USER_MODEL = 'rh.User'
+AUTH_USER_MODEL = 'accounts.Account'
 
 # Internationalization
 # https://docs.djangoproject.com/en/4.0/topics/i18n/
@@ -142,4 +144,4 @@ EMAIL_HOST_USER = env('EMAIL_HOST_USER')
 EMAIL_HOST_PASSWORD = env('EMAIL_HOST_PASSWORD')
 EMAIL_PORT = env('EMAIL_PORT')  # 2525
 
-# PLAY_DB = env("PLAY_DB")
+PLAY_DB = env("PLAY_DB")
