--- templates/stocks/stock_location_reports.html
@@ -1,94 +1,0 @@
-{% extends '_base.html' %}
-{% load crispy_forms_tags %}
-
-
-{% block content %}
-
-<div class="container px-4 px-lg-5 h-100">
-    <div class="row h-100 align-items-center justify-content-center text-left">
-        <div class="col-lg-9 align-self-end">
-            <h1 class="mb-4 mt-4">Report Hub</h1>
-            <div class="card">
-                <div class="card-body">
-                    <h3 class="card-title bg-light p-3">
-                        <i class="fa fa-calendar" aria-hidden="true"></i> &nbsp; {{report.created_at|date:"F, Y"}}
-                    </h3>
-                    {% if messages %}
-                        {% for message in messages %}
-                            {% if message.tags in 'info,warning,success' %}
-                                <div class="alert alert-{{ message.tags }}" role="alert">{{ message }}</div>
-                            {% elif message.tags == 'error' %}
-                                <div class="alert alert-danger" role="alert">{{ message }}</div>
-                            {% else %}
-                                <div class="alert alert-warning" role="alert">{{ message }}</div>
-                            {% endif %}
-                        {% endfor %}
-                    {% endif %}
-                    <h5 class="p-3">
-                        Selected Locations 
-                        <span class="badge bg-success">
-                            {{warehouse_location_stocks|length}}
-                        </span>
-                    </h5>
-                    {% for warehouse_location, formset in warehouse_location_stocks.items %}
-                        <form method="POST" enctype="multipart/form-data" class="mt-3 p-2">
-                            {% csrf_token %}
-                            <h5>
-                                <i class="fa fa-map-marker" aria-hidden="true"></i> &nbsp;{{warehouse_location.province.name}}, {{warehouse_location.district.name}}, {{warehouse_location.name}}
-                            </h5>
-                            {{ formset.management_form }}
-                            {% for form in formset %}
-                                {{ form.id }}
-                                <div class="accordion mb-3" id="stock_details_{{warehouse_location.id}}_{{forloop.counter0}}">
-                                    <div class="accordion-item">
-                                        <h2 class="accordion-header" id="stock_details_header_{{warehouse_location.id}}_{{forloop.counter0}}">
-                                            <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapse_{{warehouse_location.id}}_{{forloop.counter0}}" aria-expanded="false" aria-controls="collapse_{{warehouse_location.id}}_{{forloop.counter0}}">
-                                                {% if form.instance.id %}
-                                                    <i class="fa fa-list" aria-hidden="true"></i> &nbsp; 
-                                                    {{ form.instance.stock_type }}, {{ form.instance.targeted_groups }}
-                                                {% else %}
-                                                    <i class="fa fa-plus-square"></i>&nbsp; Add Stock 
-                                                {% endif %}
-                                            </button>
-                                        </h2>
-                                        <div id="collapse_{{warehouse_location.id}}_{{forloop.counter0}}" class="accordion-collapse collapse hide" aria-labelledby="stock_details_header_{{warehouse_location.id}}_{{forloop.counter0}}" data-bs-parent="#stock_details_{{warehouse_location.id}}_{{forloop.counter0}}">
-                                            <div class="accordion-body" id="accordion_body_{{warehouse_location.id}}_{{forloop.counter0}}">
-                                                <div class="row">
-                                                    <div class="col">
-                                                    {{ form.cluster|as_crispy_field}}
-                                                    {{ form.stock_type|as_crispy_field}}
-                                                    {{ form.qty_in_stock|as_crispy_field}}
-                                                    </div>
-                                                    <div class="col">
-                                                    {{ form.stock_purpose|as_crispy_field}}
-                                                    {{ form.status|as_crispy_field}}
-                                                    {{ form.qty_in_pipeline|as_crispy_field}}
-                                                    </div>
-                                                    <div class="col">
-                                                    {{ form.targeted_groups|as_crispy_field}}
-                                                    {{ form.stock_unit|as_crispy_field}}
-                                                    {{ form.beneficiary_coverage|as_crispy_field}}
-                                                    </div>
-                                                </div>
-                                                
-                                                <div class="mb-3 mt-3 text-end">
-                                                    <input type="submit" class="btn btn-outline-primary" name="create_{{ form.instance.id }}_stock_detail" value="Save"/>
-                                                    {% if form.instance.id %}
-                                                        <input type="submit" class="btn btn-outline-danger" name="delete_{{ form.instance.id }}_stock_detail" value="Delete"/>
-                                                    {% endif %} 
-                                                </div>
-                                                
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>  
-                            {% endfor %}
-                        </form>
-                    {% endfor %}
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-{% endblock %}
