+++ urls.py
@@ -18,6 +18,8 @@ from django.urls import path, include
 
 urlpatterns = [
     path('', include('rh.urls')),
+    path('', include('stock.urls')),
+    path('', include('accounts.urls')),
     path('', include('django.contrib.auth.urls')),
     path('admin/', admin.site.urls),
 ]
